@@ -1,13 +1,23 @@
 {
   "name": "snarkjs",
-  "version": "0.1.20",
+  "type": "module",
+  "version": "0.3.23",
   "description": "zkSNARKs implementation in JavaScript",
-  "main": "index.js",
+  "main": "./build/main.cjs",
+  "module": "./main.js",
+  "exports": {
+    "import": "./main.js",
+    "require": "./build/main.cjs"
+  },
   "scripts": {
-    "test": "mocha"
+    "test": "mocha",
+    "build": "rollup -c config/rollup.cjs.config.js",
+    "buildcli": "rollup -c config/rollup.cli.config.js",
+    "buildiife": "BROWSER=true rollup -c config/rollup.iife.config.js",
+    "buildiifemin": "BROWSER=true rollup -c config/rollup.iife_min.config.js"
   },
   "bin": {
-    "snarkjs": "cli.js"
+    "snarkjs": "build/cli.cjs"
   },
   "directories": {
     "templates": "templates"
@@ -28,15 +38,23 @@
     "url": "https://github.com/iden3/snarkjs.git"
   },
   "dependencies": {
-    "big-integer": "^1.6.43",
-    "chai": "^4.2.0",
-    "escape-string-regexp": "^1.0.5",
-    "eslint": "^5.16.0",
-    "keccak": "^2.0.0",
-    "yargs": "^12.0.5"
+    "blake2b-wasm": "https://github.com/jbaylina/blake2b-wasm.git",
+    "circom_runtime": "0.1.4",
+    "fastfile": "0.0.13",
+    "ffjavascript": "0.2.10",
+    "logplease": "^1.2.15",
+    "r1csfile": "0.0.12"
   },
   "devDependencies": {
-    "eslint-plugin-mocha": "^5.3.0",
-    "mocha": "^5.2.0"
+    "chai": "^4.2.0",
+    "eslint": "^6.8.0",
+    "mocha": "^7.1.1",
+    "rollup": "^2.20.0",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-ignore": "^1.0.6",
+    "rollup-plugin-json": "^4.0.0",
+    "rollup-plugin-node-resolve": "^5.2.0",
+    "rollup-plugin-replace": "^2.2.0",
+    "rollup-plugin-terser": "^6.1.0"
   }
 }
